agm: plugins: Update errno in amp_pcm_get_param_get
Update errno in amp_pcm_get_param_get to avoid false update of SSR down
in agm clients. When sound card is up after SSR, and this API is called,
and it fails because of some other error, errno is still -ENETRESET due
to some other APIs failure during SSR resulting in false report of SSR
down in agm clients.

Change-Id: If3b73a3b2a06c12108b8ec1456ad34fd28b1abc9